MONGODB/etl/requirements.txt
pymongo>=4.6.0
pandas>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
mysql-connector-python>=8.2.0
pyodbc>=5.0.0
neo4j>=5.14.0
